Some cleanups, and experiments.
