+++ server.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python3
 
 import argparse
-import jsonpickle
 import json
 import os
 import random
@@ -29,6 +28,8 @@ from src.profession import ProfessionManager, Profession
 from src.monster import MonsterManager
 from src.worldmap import Worldmap
 from src.passhash import makeSalt
+from src.serializer import encode_packet, decode_packet
+
 
 
 class OverMap:  # when the character pulls up the OverMap. a OverMap for each character will have to be stored for undiscovered areas and when they use maps.
@@ -217,7 +218,7 @@ class Server(MastermindServerTCP):
         )
 
         with open(path, "w") as fp:
-            _encoded = jsonpickle.encode(character, unpicklable=False, warn=True)
+            _encoded = encode_packet(character, unpicklable=False, warn=True)
             json.dump(_encoded, fp)
 
         self._log.info(
@@ -313,11 +314,11 @@ class Server(MastermindServerTCP):
                 self.localmaps[data['args'][0]] = self.worldmap.get_chunks_near_position(
                     self.characters[data['args'][0]].position
                 )
-                self.callback_client_send(connection_object, jsonpickle.encode(self.localmaps[data['args'][0]]))
+                self.callback_client_send(connection_object, encode_packet(self.localmaps[data['args'][0]]))
 
             if _command["command"] == "completed_character":
                 if not data["ident"] in self.characters:
-                    _character = jsonpickle.decode(data["args"][0])
+                    _character = decode_packet(data["args"][0])
                     # this character doesn't exist in the world yet.
                     self.handle_new_character(data["ident"], _character)
                     self._log.debug(
@@ -349,7 +350,7 @@ class Server(MastermindServerTCP):
                     self.characters[data["args"][0]].position
                 )
                 self.callback_client_send(
-                    connection_object, jsonpickle.encode(self.localmaps[data["args"][0]])
+                    connection_object, encode_packet(self.localmaps[data["args"][0]])
                 )
 
             # all the commands that are actions need to be put into the command_queue then we will loop through the queue each turn and process the actions.
